todo_service/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
pydantic==2.8.2
PyJWT==2.9.0
passlib==1.7.4
email-validator==2.2.0
python-dotenv==1.0.1
alembic==1.13.2
sqlmodel==0.0.22
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-cov==5.0.0

